requirements.txt
amqp==2.2.2
billiard==3.5.0.3
boto3==1.9.14
botocore==1.12.14
cachetools==2.1.0
celery==4.1.0
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.3
django-allauth==0.38.0
django-heroku==0.3.1
django-jinja==2.4.1
django-rest-auth==0.9.3
django-storages==1.7.1
djangorestframework==3.8.2
docutils==0.14
et-xmlfile==1.0.1
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.12.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.7.1
idna==2.7
jdcal==1.3
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
numpy==1.14.2
oauthlib==2.1.0
openpyxl==2.5.1
pandas==0.22.0
protobuf==3.6.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
urlwait==0.4
vine==1.1.4
whitenoise==3.3.1
xlrd==1.1.0
